FB Data: Updated script to read from existing files & handle empty responses

+++ FB-Data/get_posts.py
@@ -1,4 +1,6 @@
+import os
 import json
+import time
 
 import facebook
 import requests
@@ -12,16 +14,39 @@ posts = graph.get_connections(profile['id'], 'posts')
 count = 1
 while True:
 
-    # Write JSON to Disk
     filename = "Data/Posts/%d.json" % count
-    with open(filename, "w") as out:
-        json.dump(posts, out)
-        print("Written data to: %s" % filename)
+
+    # Write JSON to Disk, if it doesn't already exist.
+    if not os.path.isfile(filename):
+
+        with open(filename, "w") as out:
+            json.dump(posts, out)
+            print("Written data to: %s" % filename)
+
+        nxt = posts['paging']['next']
+        posts = requests.get(nxt).json()
+
+    else:
+
+        with open(filename) as inp:
+            posts = json.load(inp)
+            print("Loaded data from: %s" % filename)
 
     count += 1
 
-    try:
-        posts = requests.get(posts['paging']['next']).json()
-    except KeyError:
-        print("The End.")
-        break
+    # There are two reasons for an empty response:
+    #
+    # 1. We got rate limited.
+    # 2. This was actually the last post.
+    if not posts["data"]:
+        for interval in (15, 30, 60, 100):
+            posts = requests.get(nxt).json()
+
+            if not posts["data"]:
+                print("Got an empty Response. Retrying in %d seconds." % interval)
+                time.sleep(interval)
+            else:
+                break
+        else:
+            print("The End.")
+            break
